# server/api/openapi.yaml
openapi: 3.0.3
info:
  title: Pantheon Spins API
  version: 0.1.0
  description: |
    Social casino API. All currency is play money ("Coins"): it cannot be bought
    with real money and cannot be withdrawn.

    Every game outcome is computed on the server. The client sends only a game
    ID and a bet, then animates the returned result.

    State-changing requests (POST/PATCH/PUT/DELETE) must send the `X-CSRF-Token`
    header with the value of the `ps_csrf` cookie, which the server sets on every
    response that lacks one.
servers:
  - url: /api/v1
security:
  - cookieAuth: []
...
paths:
...
  /auth/password-reset/request:
    post:
      tags: [auth]
      operationId: requestPasswordReset
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetRequest"
      responses:
        "204":
          description: |
            Always returned, whether or not the email exists, so the endpoint
            cannot be used to discover accounts.
        "400": { $ref: "#/components/responses/BadRequest" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /auth/password-reset/confirm:
    post:
      tags: [auth]
      operationId: confirmPasswordReset
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetConfirm"
      responses:
        "204":
          description: Password changed. All existing sessions are revoked.
        "400": { $ref: "#/components/responses/BadRequest" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
...
  /auth/google/start:
    get:
      tags: [auth]
      operationId: googleStart
      security: []
      responses:
        "302":
          description: Redirect to Google
        "404": { $ref: "#/components/responses/NotFound" }
  /auth/google/callback:
    get:
      tags: [auth]
      operationId: googleCallback
      security: []
      parameters:
        - { name: code, in: query, schema: { type: string } }
        - { name: state, in: query, schema: { type: string } }
        - { name: error, in: query, schema: { type: string } }
      responses:
        "302":
          description: Redirect back to the web app
  /me:
    get:
      tags: [profile]
      operationId: getMe
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
        "401": { $ref: "#/components/responses/Unauthorized" }
    patch:
      tags: [profile]
      operationId: updateMe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
  /wallet:
    get:
      tags: [wallet]
      operationId: getWallet
      responses:
        "200":
          description: Wallet balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
        "401": { $ref: "#/components/responses/Unauthorized" }
  /wallet/transactions:
    get:
      tags: [wallet]
      operationId: listTransactions
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/BeforeId"
      responses:
        "200":
          description: Newest first
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionPage"
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /games/{gameId}/bonus:
    get:
      tags: [games]
      operationId: getActiveBonus
      description: Returns the unresolved bonus on this game so the client can resume it.
      parameters:
        - $ref: "#/components/parameters/GameId"
      responses:
        "200":
          description: Active bonus
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BonusState"
        "204":
          description: No active bonus
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /games/{gameId}/bonus/actions:
    post:
      tags: [games]
      operationId: bonusAction
      parameters:
        - $ref: "#/components/parameters/GameId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BonusActionRequest"
      responses:
        "200":
          description: The step outcome and the new bonus state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BonusStepResult"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /spins:
    get:
      tags: [history]
      operationId: listSpins
      parameters:
        - $ref: "#/components/parameters/Limit"
        - name: before
          in: query
          description: Return spins created before this spin ID (cursor)
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Newest first
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpinHistoryPage"
        "401": { $ref: "#/components/responses/Unauthorized" }
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: ps_session
  parameters:
    GameId:
      name: gameId
      in: path
      required: true
      schema:
        type: string
        pattern: "^[a-z0-9-]{1,64}$"
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    BeforeId:
      name: before
      in: query
      description: Return items with an ID lower than this (cursor)
      schema: { type: integer, format: int64 }
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Not signed in, or bad credentials
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: Conflicts with current state
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    TooManyRequests:
      description: Rate limited
      headers:
        Retry-After:
          schema: { type: integer }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code, e.g. `insufficient_funds`
        message:
          type: string
...
    PasswordResetRequest:
      type: object
      required: [email]
      properties:
        email: { type: string, maxLength: 254 }
    PasswordResetConfirm:
      type: object
      required: [token, password]
      properties:
        token: { type: string, maxLength: 128 }
        password: { type: string, minLength: 10, maxLength: 128 }
    Username:
      type: string
      minLength: 3
      maxLength: 20
      pattern: "^[A-Za-z0-9_]+$"
    AvatarId:
      type: string
      description: One of the preset avatar IDs
      enum: [raven, owl, serpent, scarab, wolf, phoenix, kitsune, spider, fox, comet]
    Me:
      type: object
      required: [id, email, username, avatar, balance, createdAt, hasPassword]
      properties:
        id: { type: string, format: uuid }
        email: { type: string }
        username: { $ref: "#/components/schemas/Username" }
        avatar: { $ref: "#/components/schemas/AvatarId" }
        balance: { $ref: "#/components/schemas/Coins" }
        hasPassword: { type: boolean }
        createdAt: { type: string, format: date-time }
    UpdateProfileRequest:
      type: object
      properties:
        username: { $ref: "#/components/schemas/Username" }
        avatar: { $ref: "#/components/schemas/AvatarId" }
    Coins:
      type: integer
      format: int64
      description: Play-money amount. Always an integer number of Coins.
    Wallet:
      type: object
      required: [balance]
      properties:
        balance: { $ref: "#/components/schemas/Coins" }
    TransactionKind:
      type: string
      enum: [signup_bonus, spin_bet, spin_win, bonus_win, daily_bonus, hourly_bonus, refill]
    Transaction:
      type: object
      required: [id, kind, amount, balanceAfter, createdAt]
      properties:
        id: { type: integer, format: int64 }
        kind: { $ref: "#/components/schemas/TransactionKind" }
        amount:
          type: integer
          format: int64
          description: Signed change in balance
        balanceAfter: { $ref: "#/components/schemas/Coins" }
        gameId: { type: string }
        createdAt: { type: string, format: date-time }
    TransactionPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Transaction" }
        nextBefore:
          type: integer
          format: int64
          description: Pass as `before` to get the next page. Absent on the last page.

    # ---------- Games ----------
...
    Position:
      type: object
      required: [reel, row]
      properties:
        reel: { type: integer }
        row: { type: integer }
    Grid:
      type: array
      description: |
        Symbol IDs by reel (column), each from top to bottom. Reels may differ in
        length (Megaways). Hex grids use offset columns: `reel` is the column and
        `row` the cell within it.
      items:
        type: array
        items: { type: string }
    Win:
      type: object
      required: [symbol, positions, amount]
      properties:
        symbol: { type: string }
        positions:
          type: array
          items: { $ref: "#/components/schemas/Position" }
        amount: { $ref: "#/components/schemas/Coins" }
        multiplier:
          type: integer
          description: Total multiplier applied to this win (1 when none)
        ways: { type: integer, description: Ways-pay games only }
        line: { type: integer, description: Line-pay games only }
        count:
          type: integer
          description: Symbols matched (reels for ways/lines, cluster size for clusters)
    GameEvent:
      type: object
      description: |
        A visual event the client animates in order, e.g. `wild_added`,
        `multiplier`, `symbol_transform`, `serpent_move`, `lava_spread`.
        `positions` is what the event affects; `data` holds event-specific
        values whose shape each game documents.
      required: [type]
      properties:
        type: { type: string }
        positions:
          type: array
          items: { $ref: "#/components/schemas/Position" }
        symbol: { type: string }
        value: { type: integer, format: int64 }
        data:
          type: object
          additionalProperties: true
    SpinStep:
      type: object
      description: |
        One grid state. A spin without cascades has one step. For cascades,
        each step's `removed` positions are cleared and `grid` of the next step
        is the board after symbols drop in.
      required: [grid, wins, stepWin]
      properties:
        grid: { $ref: "#/components/schemas/Grid" }
        events:
          type: array
          description: Events to play before evaluating this step's wins
          items: { $ref: "#/components/schemas/GameEvent" }
        wins:
          type: array
          items: { $ref: "#/components/schemas/Win" }
        removed:
          type: array
          items: { $ref: "#/components/schemas/Position" }
        multiplier:
          type: integer
          description: Step multiplier (cascade multipliers etc.), 1 when none
        stepWin: { $ref: "#/components/schemas/Coins" }
...
    SpinOutcome:
      type: object
      description: The animated part of a spin; reused by free-spin bonus steps.
      required: [steps, totalWin]
      properties:
        steps:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/SpinStep" }
        events:
          type: array
          description: Events after all steps (e.g. end-of-spin multipliers)
          items: { $ref: "#/components/schemas/GameEvent" }
        totalWin: { $ref: "#/components/schemas/Coins" }
        anticipation:
          type: array
          description: Reels to spin slower because a feature may land
          items: { type: integer }
...
    BonusKind:
      type: string
      enum: [free_spins, pick, ladder, wheel, plinko, trail, hold_and_win, timing, route]
    BonusStatus:
      type: string
      enum: [active, completed]
    BonusState:
      type: object
      description: |
        The public part of a bonus state machine. Hidden values (e.g. what an
        unpicked tile contains) never leave the server.
      required: [bonusId, gameId, kind, status, bet, totalWin, actions, data, step]
      properties:
        bonusId: { type: string, format: uuid }
        gameId: { type: string }
        kind: { $ref: "#/components/schemas/BonusKind" }
        status: { $ref: "#/components/schemas/BonusStatus" }
        bet: { $ref: "#/components/schemas/Coins" }
        totalWin:
          $ref: "#/components/schemas/Coins"
        step:
          type: integer
          description: Number of actions taken so far
        actions:
          type: array
          description: Actions the player may send next, e.g. `spin`, `pick`, `choose`
          items: { $ref: "#/components/schemas/BonusActionOption" }
        data:
          type: object
          description: Game-specific public state (spins left, picked tiles, ladder rung…)
          additionalProperties: true
    BonusActionOption:
      type: object
      required: [action]
      properties:
        action: { type: string }
        choices:
          type: array
          description: Valid values for `choice`, when the action takes one
          items: { type: string }
    BonusActionRequest:
      type: object
      required: [action, step]
      properties:
        action: { type: string, maxLength: 32 }
        choice: { type: string, maxLength: 64 }
        step:
          type: integer
          description: |
            The `step` of the state the client acted on. A mismatch returns 409,
            which makes double-clicks and retries safe.
    BonusStepResult:
      type: object
      required: [state, stepWin, balance]
      properties:
        state: { $ref: "#/components/schemas/BonusState" }
        stepWin: { $ref: "#/components/schemas/Coins" }
        balance: { $ref: "#/components/schemas/Coins" }
        outcome:
          $ref: "#/components/schemas/SpinOutcome"
        reveal:
          type: object
          description: What this action revealed (pick contents, wheel segment, …)
          additionalProperties: true
    SpinHistoryItem:
      type: object
      required: [spinId, gameId, bet, win, createdAt]
      properties:
        spinId: { type: string, format: uuid }
        gameId: { type: string }
        bet: { $ref: "#/components/schemas/Coins" }
        win: { $ref: "#/components/schemas/Coins" }
        bonusTriggered: { type: boolean }
        createdAt: { type: string, format: date-time }
    SpinHistoryPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/SpinHistoryItem" }
        nextBefore: { type: string, format: uuid }